+++ engine/include/engine/initialize.h
@@ -8,7 +8,7 @@
 #include "common.h"
 #include "execution_graph/logic_controllers/CacheMachine.h"
 
-std::pair<std::pair<std::shared_ptr<ral::cache::CacheMachine>,std::shared_ptr<ral::cache::CacheMachine> >, int> initialize(int ralId,
+std::pair<std::pair<std::shared_ptr<ral::cache::CacheMachine>,std::shared_ptr<ral::cache::CacheMachine> >, int> initialize(uint16_t ralId,
 	std::string worker_id,
 	int gpuId,
 	std::string network_iface_name,
@@ -26,7 +26,7 @@ void finalize();
 size_t getFreeMemory();
 extern "C" {
 
-error_code_t initialize_C(int ralId,
+error_code_t initialize_C(uint16_t ralId,
 	std::string worker_id,
 	int gpuId,
 	std::string network_iface_name,
